+++ Humans.java
@@ -52,7 +52,7 @@ public class Humans extends LivingSpecies{
         System.out.println(job + " is fishing.");
     }
 
-    public void pollute(Climate climate, float amount) {
+    public void pollute(Climate climate, double amount) {
         climate.incrementPollution(amount);
         System.out.println(job + " has polluted the environment by " + amount + " units.");
     }
